+++ virtual-keyboard/src/assets/js/commonNode.js
